pyproject.toml
[tool.poetry]
name = "realtime-streaming"
version = "0.1.0"
description = ""
authors = ["Vadim Titko <vtitko27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index-core = "^0.11.17"
pyaudio = "^0.2.14"
pynput = "^1.7.7"
pydub = "^0.25.1"
websockets = "^13.1"
wave = "^0.0.2"
python-dotenv = "^1.1.0"
websocket-client = "^1.8.0"
openai = "^1.76.0"
python-vlc = "^3.0.21203"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
elevenlabs = "^1.57.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
